pdfwrite - fix stream MD5 hashing
Objects in pdfwrite use an MD5 hash to check for equality. The hash is lazily
created, and stored for later reuse.

The stream object hashing wasn't adding the hash of the dictionary associated
with the stream to the hash of the stream contents. This meant that 2 streams
with the same data, but different associated dictionaries, could be wrongly
identified as equal.

In practice this condition is likely to be extremely rare, it was detected
while working on the colour code and was caused by using a hand-crafted
PostScript file which reused many components when constructing multiple
colour spaces.

No differences expected