pdfwrite - initialise structure member 'stream_md5_valid
Bugs #693759, #693760, #693761

Found by valgrind, there were conditions under which the member would not
be initialised before we tested it, initialising it to 0 ensures this can't
happen.

In practice this is unlikely to ever cause a problem, but its best to be safe
as we could conceivably attempt to access invalid memory locations if this
member isn't initialised.

No differences expected.