Valgrind: Fix spurious warnings in gc
The Ghostscript object headers have a union of bitfields; these are
arranged so that they are all of the form 1 bit, then n-1 bits
(where n is a power of 2), so they pad nicely to a nice size.

Specifically they are designed to work like this (as far as I can
make out) to allow the 'alone' bit to be preserved during gcs
when the n-1 bit fields are used. Valgrind, I think, finds this
problematic.

The fix in this commit is simply to initialise both the alone bit
and the padding bits at the same time (any even halfway decent
compiler should combine these into a single byte write, so we
should actually find this new safer code is faster than the previous
single bit initing code). This is safe because the other unions
are only used during gc.

Cluster testing shows no differences, and local valgrind runs
appear cleaner.