Bug 693824: Valgrind problems in interpolated images.
The first of the valgrind problems reported here is a false
positive; when scaling just a subsection of an image we
scale extra 'support' pixels. Due to the way this works we
end up scaling a few extra lines/columns around the rectangle
we actually need. Some of the input values to these extra
lines are not copied in, hence we see valgrind complaining
about uninitialised values.

We fix this with some PACIFY_VALGRIND code to memset the
input buffers to 0.

This commit also fixes valgrind errors seen later due to
us trying to color convert color values outside of the
active region; we fix this by restricting the area we
color convert. This will give speedups in some cases.