Fix memory corruption problem with HPGL/2 page sizes.
Setting a page size in the HPGL/2 code depended on memory in a local
stack frame.  This problem was introduced when custom page size
support was added to PCL which resulted in reorganizing the paper type
set up.  Unfortunately HPGL/2 - RTL mode was not updated to use the
new paper setup when the PCL change was made.