Fix valgrind issue, bug 693735. image_render_mono condition using UMR.
The psrc scanning loop did not stop at the end of the data in some
cases. Since data past the image would tend to be variable, this
would usually stop scanning fairly soon and the 'fit_' macro logic
in fill/copy routines would make sure to limit the painting to a
valid area. Strange that this didn't show up in image artifacts.