pdfwrite - correct tracking of current point with word spcaing
Bug #693285 "gs adds spaces and breaks formatting of original pdf file"

This one is quite subtle and caused by a number of interacting issues.

Firstly the design grid conversion of TrueType fonts can lead to TT CIDFonts
ending up with very slightly different widths from the font to the widths in
the /W array (rounding errors). When emitting text we then use the TJ operator
and emit tiny movements for each glyph.

The second part then, is the limit we have of 50 movements per TJ operation.
There's no indication why we have this limit but I assume its some ancient
version of Acrobat. When we get to 50 moves we close the current line of text
and begin a new one.

The third part is that, in order to start the new line from the correct x offset
(so it lines up properly with the previous line), we keep a running track of
the current position and use that to tell us where to start the new line.
This value was getting the word spacing value applied to it whenever the
current character was a space (0x20). However, we only apply the word spacing
to *single* byte encodings in PDF (unlike PostScript) and the font in this
particular file doesn't have a single byte range.

To fix this we need to know if the text came from a (a)widthshow operation
from PDF, or from PostScript. Fortunately the text enumerator already has
a 'single_byte_space' member which tells us this. If it was from PDF we then
need to check if the character encoded used 1 byte or more than one, and
not apply the word spacing if it used more than 1. To this end we now pass
in the number of bytes to process_text_modify_width.

This change exhibits progressions in fts_20_2001.pdf and fts_22_2202.pdf