pdfwrite - prevent rounding errors in TrueType widths
See Bug #693825

Rounding errors converting TrueType widths into PostScript/PDF led to us
emitting lots of tiny glyph movements. These were visually imperceptible
but led to larger than necessary output, because of all the tiny shifts.

This commit fixes the widths from TrueType fonts to the nearest integer
which makes it much more likely they will match the entries in the /W or
/Widths arrays, and thus not result in tiny movements.

The file in Bug #698325 is 5% smaller with this change, and the output PDF
file does actually match the original more closely (when using a high zoom
factor and careful inspection)

This does (unsurprisingly) cause a lot of differences in the test suite,
but I've looked at all of them and can see no definite problems. The
differences are pretty much single pixel shifts and this can sometimes
cause low resolution output to be slightly better, or slightly worse.