Add BGPrint prn device parameter to render clist pages while parsing.
This enhancement closes the clist files of the page, then starts a
printing thread (prn_print_page_in_background) to write out the page
and returns to allow parsing of the next page into a new set of clist
files while the 'bg printing' thread writes out the data, performing
compression and other conversion that the BGPrint=false mode performs
in the main (parsing) thread.

If NumRenderingThreads > 0, the band rendering will also be performed
for multiple bands in separate threads.

Thus, the parsing of pages after the first will potentially overlap
the rendering and writing of the previous page. No queue is maintained,
so the overlap is limited to the next page.

If bg_print.device is not NULL, then we have a page that was rendered
in the background that we may need to wait for before pprinting the next
page or exiting. After we call the prn_finish_bg_print, we may need to
re-open the printer in case it was closed (most commonly due to %d in
the fname).

When printing in foreground mode, the bg_print.device will be NULL.

Note, We can't use the icc_table created for the foreground since it may
be in GC'd memory (it must be created with the thread's allocator).

The equivalent_cmyk_colors are needed by the backgorund print device.
Add .EquivCMYKColors device parameter which is an array of 5 integers per
spot color. The first integer is 'valid' followed by the 'frac' C, M, Y
and K values. This allows the gs_putdeviceparams to have the spot color
information.

The foreground device was relying on the print_page procedure to free
the separation names from a page before starting the next page. Move
this to gdev_prn_output_page instead. Also since this is common, we
don't need it in gdevpsd and gdevtsep print_page functions.

Expected difference when BGPrint=true is 99-01-fixed.PS for all devices
due to adding the .EquivCMYKColors device parameter.