ps2write - do not emit CIEBased spaces as ICCBased for PostScript
This was an oversight when I was trying to preserve CIEBased colour spaces,
it turns out that its not possible to preserve them, and I accidentally left
the code emitting them as ICCBased, which doesn't work in PostScript.

We now fall back to the old approach, in future we hope to use LCMS2 to
create a CIEBased space from the ICCBased profile for us.

No differences expected, code not enabled by default