Bug 693843: Add a gs_sprintf() function, and supporting code
Same as sprintf() but certain to be unaffected by localization.

Utilizes the "trio" string library, also added here. There is one modification
to trio, in trio.c and labelled with "/* GS Modification */" - by default, trio
is a little aggressive in the point it switches to exponent notation, and the
patch brings is more line with what we need/expect.

trio also includes a sscanf() function which we can utilise if required.

This causes rounding differences in the pdfwrite/ps2write output in a number
of the cluster files. They all appear to be pixel variations.