Initial commit of adding neutral color monitoring
This commit adds in the command line option -dGrayDetection=true/false.   When set, the special CMM procedures
will be inserted to monitor the presence of non-neutral colors.   When a non-neutral color is encountered
no additional monitoring will be preformed.  Monitoring is done during the clist writing phase and as such
high level images are also monitored.   CMYK Colors are consider non neutral if C=M=Y=0 (or very close).  RGB
colors are considered neutral if R=G=B or very close and CIELAB colors are considered neutral if a=128 and b=128
or very close.   A fuzzy range can be set for how close we need to be via defines in the code.  The
current values were set based upon using images that were created to be monochrome in Photoshop using ICC
profiles.   Testing for this was performed for a variety of image types.  The remaining part of this project,
which relates to bug number 693583 is to do something useful with the presence of monochrome only data during
the clist playback phase.   The values pageneutralcolor = true and graydetection, which are member variables in
cmm_dev_profile_t, during clist playback will indicate that the page is neutral.  Note that the clist playback
device should set pageneutralcolor = false during the actual playback to avoid continued monitoring of
high level image data, which would occur as the data is pushed through the CMM.  graydetection is the
device parameter and as such,  pageneutralcolor only has meaning when graydetection = true.