Have gdev_prn_open_printer_seekable() work sensibly
If gdev_prn_open_printer_seekable() gets called with the seekable flag set,
and it can't return a seekable file object, return an error, rather than
silently ignore the flag.

Add a generic gdev_prn_output_page_positionable() function to replace the one
that was dedicated to the TIFF devices, and update the TIFF devices and the
two other devices which require positionable output files to use it.

No cluster differences.