Allow devices to inform the PCL interpreter that it must not alter palettes
Bug #692786 "Problem with HPGL paletted colour"

pdfwrite (and potentially other devices) can't cope with situations where the
interpreter alters the 'palette' (the lookup table in a /Indexed space) after
the space has been set to be the current color space.

However, if we unilaterally preserve the palette, it seems that this causes
performance problems (see bug #692786.

Here we add a new special op 'needs_invariant_palette' which allows a device
to communicate this. We alter the PCL interpreter to inquire the device's
requirements, and aler pdfwrite to communicate its needs.

This results in > 300 differences with pdfwrite, all of them progressions