Add high level pattern support to PCL and HPGL
Unlike the PXL interpreter, but like XPS, the coding has been done in the
implementation routines rather than at the interpreter level. There are two
reasons for this; firstly the PCL interpreter already called gx_dev_set_color
and secondly the HPGL interpreter makes extensive use of macros which weren't
readily amenable to being altered in a way which would have allowed the
gs_error_Remap_Color error to be propogated upwards as required.

There are some differences with this commit, they are either very minor or
progressions (in my opinion).