pdfwrite - fix buffer overrun in CIDToGIDMap
Bug #693854 "PDFA/2b conversion crashes often"

When converting TrueType fonts into CIDFOnts, because we are creating a PDF/A
output file, the calculated length of the CIDToGIDMap could be out by one
as the first CID is 0.

This later led to us writing beyond the end of the buffer, which could cause
real problems, depending on what was overwritten.

No differences expected.