Add gdev_prn_bg_output_page (and _seekable) to support BGPrint.
Some devices are not compatible with background printing, if they are
not thread safe (use globals) or update the device structure.
Simple updates to the device structure can be handled in a device's
own output_page function that performs any actions needed before
calling gdev_prn_bg_output_page, then performing any action needed
after the page is printed (or started in the background). The cups
device is an example of this.

If the page is printed in the background, it uses a clone of the device
structure, so any changes to that structure will not be reflected in
the foreground device.

The devices were reviewed and those that were changed to use gdev_prn_bg
output_page proc should be safe for background printing, but only
those that are used in regression testing have been tested by forcing
BGPrint=true.