pdfwrite - handle Intellifonts as type 3 fonts
In order to deal with high level patterns used as the fill colour for an
Intellifont, we need to handle Intellifonts as type 3 fonts, just like
the stick font, and PCL bitmap fonts.

This results in numerous differences; in general text is bolder at
low resolution, the file 23-15.bin exposes a probable GS bug where the
glyphs are clipped at higher resolutions.

However, text can now be searched (also copy/paste), and the PDF files are
potentially smaller due to not having to emit each glyph description every
time it is used.