Finish GrayDetection implementation and add pnmcmyk example device.
The pnmcmyk device will output a P5 pgm file if the page source colors
are all (near) neutral, otherwise will output PAM P7 32-bit CMYK.
This device is intended to be an example for a customer.

Note that there is a conditional compile based on PESSIMISTIC_CMYK_NEUTRAL
that, if defined, will assume that if any C, M, or Y component has a value
greater than DEVICE_NEUTRAL the page will be treated as color (even if all
three components are 100%). This may be what the customer requested.
The default comparison is based on the C, M, and Y components be close to
each other (similar to RGB).

Also disable graydetection during clist playback. We don't need worry about
page mode because colors are already processed as the image is rendered
into the page buffer, but high level images in the clist would slow down
if graydetection was performed again.