pdfwrite - DO not leave dangling references in font/matrix cache
Bug #693948 "%d not working with pcl"

Under some conditions (TrueType fonts, but not CID TT), we end up adding the
font to the font/matrix cache. However, the font we add is a copy of the
original.

If we are using %d to emit multiple files, then we free our copies after
every page. However, we did not purge the cache, leading to stale pointers
in it. If we then reused a font we could end up trying to use a stale
pointer with unfortunate results.

This commit adds a 'notify' procedure so that we clear copied fonts from
the cache when they are released.

This caused one difference in the regression tests, which I do not understand
but is not visilby different....

Signed-off-by: Ken Sharp <ken.sharp@artifex.com>