vector device - ensure colour changes are not restored away by clip
Bug #693951 "pswrite/epswrite is changing colors when creating an eps or an ps file from a ps file"

When an image is degenerated into trapezoids, then when writing the trapezoids
we first set the colour, then ensure any clip path is satisfied, then emit
the trapezoid and fill it.

However, if a clip path *is* pending, then it will emit a 'grestore' which
restores away the colour we initially set. This commit ensures that if a
clip is pending, we set the colour again after completing the clip.

No differences expected.