Bug 688949: allow Ghostscript to build outside src tree
These changes allow you to run the Ghostscript configure script and make from
outside the source tree. Allowing the use of a "build directory" so the
source directory stays unadulterated.

The cluster test through up a load of compiler warnings, but none of them are
new, as far as I can tell.

No cluster differences.