Squash some warnings.
In gs/base/gsroprun1.h we cast a pointer to an int and then & 3
to find alignment. Given ints are always > 2 bits, this is always
safe, but the compiler whinges anyway. We'd like to use ptrdiff_t
instead to avoid the whinging, but it's not defined. So define
our own using the ARCH_ macros.

In gs/base/gxclrect.c, we check for the size of a type, and then
shift down by it. The compiler complains that the shift may be
too large to be defined in C, not realising that this is precisely
what we have just checked. We work around this by shifting in 2
stages.