Add PSFitPage and FitPage options to allow fitting PS files to a page.
The FitPage option is added as a convenience to set all three *FitPage
options, so -dFitPage is the same as -dEPSFitPage -dPDFFitPage -dPSFitPage.
The -dPSFitPage option is new and works by establishing the default
Polcies PageSize to 13.

When the Policy is 3 or 13, the page size matching logic rotates, scales,
and center a PostScript file's requested page size on the best fit medium.
When the Policy is 13, the the match_page_size enables the scaling down
(as with Policy 3) as well as up.

Only populate the InputAttr dictionary with the current page size if
FIXEDMEDIA is set (explicitly, or impiled by -g) so that only that
page size is available for match_page_size.

When the policy is 13, we enable the PageSize in the setpagedevice
merged request even if FIXEDMEDIA is set.

The make_adjustment_matrix has the check for factor < 1 removed since
this is accounted for in the caller setting the 'scale' boolean (larger).

The documentation in doc/Use.htm is updated for both the -dPSFitPage
option and the -dFitPage option.