epswrite - flush stream before writing page trailer
Bug #694061 "epswrite generates bad file"

The code (quite deliberately) writes the page trailer *directly to the file*
before flushing the stream. This resulted in the page trailer being emitted
part way through a path construction and fill. The fill is completed by the
flushing of the stream.

In this commit we flush the stream first.

No differences expected, the cluster does not test epswrite.