pdfwrite - don't cache high level pattern filled glyphs
Bug #693946 "High level pattern commit causes a regression"

If we use the cache when a glyph is filled with a high level pattern we will
get incorrect rendering, as the pattern was not passed down to the low level
routines. So make sure we don't cache glyphs under these conditions, nor
consult the cache for existing cached glyphs.

Expected Differences
14-07.bin and 19-10.bin should now work correctly