pdfwrite - when using %d in OutputFile, empty glyphs from cache
Bug #694069

When we are using '%d' in the Output File Name, we throw away the PDF font
resource after every page, but the intperpreter glyph cache remains filled.

This is a problem for type 3 fonts (specifically their PCL equivalents), as
we won't rerun the glyph description to capture it for a type 3 font, so
the glyphs don't get captured.

To solve this we empty any existing cache entry before attempting to run the
glyph. This runs into some more awkwardness with a specific kind of broken
font (no chacracter data) in some QL test files. We work around that too.