Bug 694065: address libtiff issues with WinRT build
The defaullt libtiff I/O functions use some methods proscribed in the WinRT world.

Ghostscript does not rely on these defaults, supplying our own I/O hooks
when we invoke libtiff.

So, if we're building libtiff into GS (as in the default and preferred
configuration), we leave out those default calls, and replace them with stubs
to keep the linker happy. Currently, these are empty stubs, and apply to all
builds, which may cause issues if anyone relies on linking to our built-in
libtiff, and uses these calls. If that proves to be the case, we can easily
fill-in the stubs with appropriate code, but working through the GS I/O
calls.

This also has the advantage that any port to a weird platform, we only
need to port the Ghostscript I/O, and we don't have to worry about libtiff.

If we're using a "shared" libtiff, these stubs are not included.

Unfortunately, libtiff does not pass around a context to the memory
management functions, so we currently still have no method of using the
Ghostscript memory manager.

Update to libpng 1.6.2

libpng 1.6.x dispenses with one of the function calls (lstrlenA) not allowed
in the WinRT build.

The other (ExitProcess) we work around in the makefiles, replacing it
with a simple "exit". This should never be a problem since we actually
rely on setjmp/longjmp for libpng error handling, and libpng code checks
for that option before it will get to the "exit" call.

Re-enable TIFF and PNG devices for WinRT

Rename 'PLATFORM' to 'GSPLATFORM'

nmake (at least in VS2012) defines 'PLATFORM' for certain configurations
which interfered with our use of it, so make it more Ghostscript-centric.

No cluster differences.