Bug 693965: Fix SEGV in fuzzing test (corrupt image)
The first error given by valgrind is due to the image in this
file being a 16bpp one with a horizontal flip. This tickles
an incomplete case within the interpolation code. Not
surprising as there are more options in that code than anyone
should be expected to cope with. The code is fixed here.

This leads on to the cause of the crash; the image has a matrix
with a HUGE x shift on it. This upsets some of the floating
point calculations used to calculate the render rectangle.

This code has been hardened now to cope with such errors.