Ensure that PCL bitmap glyphs with no data still execute setcachedevice
This grew out of Bug #694069, in the course of resolving that problem it
transpired that, if there was no data available for a downloaded bitmap
character, the PCL interpreter simply returned.

In conjunction with pdfwrite, this caused an endless loop due to the character
cache being overfilled. We only check the availability of space in the
character cache in setcachedevice, so its safer to make sure that we do
call it, thereby avoiding the overfilling.

No differences expected.