pdfwrite - better support for PageLabels
Bug #693929 " Regression: Error: /undefined in --run--"

This code now properly flattens a PageLabels number tree into a single node
tree, which it converts to a string for inclusion in the output.

In addition to flattening the tree we also perform considerable error checking
on the tree construction, and we limit the number of page elements to 1000.
This is because PostScript strings are limited to 64k and we use a string to
pass the data to the device for inclusion. If we exceed 1000 elements then
we simply abort the PageLabels (and print a warning on the back channel)

No differences expected, this is not a printing eleent