pdfwrite - improve handling of Unicode names
Bug#694277 " Improve pdfmark processing to handle unusual names in named destinations"

We can't handle Unicode names in PDF, technically we should write these as
strings and enter them in the names table, but we don't have a mechanism for
that for pdfmarks.

Instead we leave them as names using the special processing in pdfwrite, but
extended with additional markers. When writing these out pdfwrite converts
the NULLs to an arbitrary string. This works adequately for the bug as
these are named destinations, but its possible that other kinds of names
might be visible in the PDF application and would give the wrong result.

So this is at best a patch.

No differences expected.