Fix for too many entries in an image Decode array reported by cust 532.
This is an out-of spec file, but as usual, since Adobe handles it, so we do too.
Allow too many by setting over_error to 0 when we load the Decode array.

This required a change to dict_float_array_param so that it will load the
requested number of elements (and return that count) when over_erro == 0.

Also clean up the indenting and get rid of the unused decode_size variable.