pdfwrite - additional error checking on named destinations
Bug #694194 "Seg faults found by fuzzing in pdfmark_DOCVIEW"

The routine for creating a named destination could return an error if the
destination page was outside the valid range (-2 in this case), but the
DOCVIEW pdfmark wasn't testing the returned value.

Added code to give a meaningful error and return a rangecheck error when this
occurs.

No differences expected.