pdfwrite - colour management, convert Separation and DeviceN base spaces
When the input is DeviceN or Separation, the color conversion strategy is
Gray, RGB or CMYK, and the alternate space is not compatible, convert the
alternate space. We do this by constructing either an exponential (Separation)
or sampled (DeviceN) function, we sample the end points of each ink,
convert the alternate space representation to the requested device space
and use those as the data points ofr the function.

The representation probably isn't that great, since it linearly interpolates
between the (very few) sample points, but it does the job adequately, and
preserves the Separation/DeviceN space while still creating an appropriate
colour space in the output.

No differences expected, this code not in use yet.