Fix for bug 694293.   We are improperly clipping to the soft mask bounding box.
If the background alpha value for the softmask is something other than 0 we cannot
clip to the soft mask bounding box as regions that are outside this area will be
using the background alpha as their softmask value.  Also, we were not accounting
for the different size of the softmask when compositing.   Progressions occurred in
three test files with this fix in addition to fixing the customer file.