Change in behavior for overprint.  Fixes bugs 694295 694296 and 694067
With this change we now do simulated overprinting of cmyk AND spot colorants
for contone cmyk devices by default.   If -dSimulateOverprint=false is specified we
will not simulate overprinting of any colors.

Also, cmyk simulation of spot colors with separation devices is not handled.
i.e. if you specify -dSimulateOverprint=true -dMaxSpots=0  -sDEVICE=tiffsep1
you will see the spot colors blended in CMYK space if overprint has been
specified in the document.

Note that not all overprint situations can be simulated accurately with this
approach.  For example, if I have a spot color that resulted in a CMYK value
of 0 0 100 0 after the alternate tint transform and then I did an overprint with
a CMYK value of 0 0 1 0, the color should still be a bright yellow but will actually
be almost no color.   The issue is that we can not distinguish this case from one
where we had first laid down a CMYK of 0 0 100 0 followed by the overprint of 0 0 1 0
which should be a light yellow.  The documentation has been updated to point this out.

I reviewed all the diffs that came up in the bmpcmp and reviewed the Ghent overprint
files.