Bug 690639: change the order we do character mapping in TTFs
Instead of combining the encoding from the PDF with the Adobe Glyph List to
get a comprehensive list of glyph name mappings, then building a character
code to GID mapping from that, and (potentially) using the TTF post table to
fill in any gaps, we now build a mapping directly from the AGL, then (possibly)
fill in any gaps from the TTF post table, and *finally* fill in any remaining
gaps using the PDF's encoding.

This handles a case where the contents of the TTF cmap table don't match the
contents of the post table.

No cluster differences.