New ram file system code
Updated from the patch submitted by Michael Slade in bug #226943. This code
removes a single Unix-ism and rewrites the memory allocations to use the
GS garbage collected memory (as is appropriate for PostScript) instead of
using the system malloc.

No differences expected since we have no files that use the ramfs.