pdfwrite - silence a Valgrind warning by initialising a variable
The memory is uninitialised because the file provokes an error, and
pdfwrite gets confused about which file it is supposed to be in, The net
result is that an MD5 hash is not generated and we end up with some memory
not being initialised.

This simply sets the memory to 0 which silences Valgrind, it makes no
attempt to improve the output.

No differences expected.