Prevent gstoraster from hanging on Ghostscript failure
Thanks to Olivier Blin for this patch.

Original discription in the bug report (bug 693336):

When gs fails and the input file is bigger than the pipe capicity
(65536 bytes on Linux), gstoraster will hang.  In this case, write()
will block because the pipe is full and no reader is emptying it.  The
gs child process becomes a zombie, and write() will not fail until its
zombie is gone, which never happens with the current code.

This patch fixes the issue by adding a child reaper, which will make
write() fail when the child dies, either because of the SIGCHLD signal
coming, or because of a broken pipe error in write().