Bug 694354: cope being unable to read a candidate font file
On Unix, things like directories can be opened as "files". In the font code,
we attempt to open a candidate font file, and if that fails, handle it. We
then read some data from the file to validate it as a suitable font file. If
the data we read reveals it to be an unsuitable file, we handle that.

We were not handling the case where we could open the "file" but could not
read from it.

No cluster differences.