PDF interpreter - handle empty cref by rebuilding it.
Bug #694342 "Error: /typecheck in --run-- reading PDF file"

The PDF interpreter didn't notice that an empty xref (the xref token
immediately followed by the trailer token) was not a valid xref. Later we
would attempt to use the empty xref, and fail.

We now check the length of 'Objects' when we find the trailer token and if
it is 0 we throw an error. This causes the PDF itnerpreter to rebuild the
xref table by rescanning the whole file.