Fix wrong GrayDetection when images have MultipleDataSources true - cust 801.
Since we already had a buffer, we unpack the samples from the planes into the
buffer using 'spread' to space them so they form a chunky pixel buffer for
row_has_color. Refactor the unpacking into the buffer out of row_has_color to
the caller where we have the planes array. Note that the buffer returned from
the 'unpack' proc may be the original data source so we save that for calling
row_has_color.