Set colors up correctly when rendering Type 3 glyph
In the 'd1' operator, if the glyph metrics are clearly wrong, we fall back to
rendering the glyph uncached. The original code failed to do the required
color setup in this path through the code.

The simplest solution, rather than repeat the same code, is to call the 'd0'
operator, after appropriate stack manipulations.

Noticed in passing working on an issue from customer 532.

No cluster differences.