Add some Encodings listed in the PLRM
Bug #688710 "Fix the Encoding resource category"

As noted int he bug thread I so not consider it worthwhile to tackle some
of the propsed work described there. However the CE Encoding is missing,
as are also the ExpertEncoding and ExpertSubsetEncoding resources.

This commit adds all three resources and the NotDefEncoding which is present
in Adobe Acrobat Distiller. We also revisit the patch (gs-cvs rev
6834) originally committed against this bug, which prevented the Subversion
source control directory being enumerated as an Encoding resource. Since we
no longer use Subversion, and the current system doesn't rely on hidden
sub-directories, this is no longer required.

No differences expected