Bug #686853 "filenameforall enumerates incompatible with Adobe"
Fix the Unix/Linux file enumeration code to correctly recurse into
directories.

This causes differences in many of the Quality Logic Postscript LL3 test suite
tests because it can now overflow the size of the "scratch" string used in the
tests and that results (correctly) in a rangecheck error from filenameforall.

Obviously, that depends on from where in the directory hierarchy the executable
is being run, and the contents of the directory structure below that.