Ensure all objects created for use during setpagdevice are in local VM
Bug #687702 "gs8.15 setpagedevice /invalidaccess"

The problem is caused by currentpagdevice returning a dictionary allocated
in the current (global) VM. We later create an OutputAttributes dictionary
which is in local VM, and store it in the dictionary returned. This fails
of course.

Since these objects are only used internally to construct various intermediate
structures during the course of a setpagedevice, the easiest solution to this
is to force the VM allocation mode to local for the duration of setpagedevice.

No differences expected.