When clamping pattern BBox to page, make sure to scan in proper direction
Bug #694385 "endless loop in pattern"

The code for clamping the pattern BBox to the page starts left and below the
page, and then scans across and up detecting if any part of the pattern is
on the page.

Unfortunately, this didn't account for negative xstep and ystep. Because
we are only scanning the area, the easiest solution is just to abs() the
values, so that we are sure to scan left-right and bottom-top.

No differences expected.