Fix bug #694355, incorrect rendering of raster image.
This problem is very similar to bug #693111, and I don't think we have
a clear handle on how the interpreter should behave when locked out
commands are received in raster mode (with RTL).  For this problem we
don't shut down graphics when receiving the render algorithm command,
the other bug involved a transparency command.  A new bug will be
opened to review both of these bugs.  Although, both fixes are
effective and safe, I'm afraid they don't address the underlying
issue.