Invalidate device refs on the operand stack when we change device
Bug #690925 "SEGV after "gsave nulldevice currentdevice grestore setdevice"

currentdevice stores a reference containing a pointer to the current device
structure on the operand stack. If we close the device, then attempt to do
'setdevice' using the stored pointer from the operand stack we can get a
seg fault if we have restored away the device, because the device structure
memory has been released.

Altered nulldevice and setdevice so that when we change devices we scour
the operand stack looking for saved references to devices. If we find any
we NULL the pointer to the device. This allows us to check for a NULL device
pointer in setdevice and raise an error instead of crashing.

No differences expected.