More work on setpagedevice and global allocation mode
The commit 073f7be1671e7fa2dbbc75984e593d2279bb05d7 solved one problem,
but introduced another. Specifically Ghostscript feiled to start when
-dSAFER was specified.

The reasons are convoluted, the background is as follows:

Adobe interpreters appear to always return a local VM writeable dictionary
from currentpagedevice (despite that fact that the spec clearly says its
read-only). WordPerfect files have, in the past at least, relied upon this
and tried to write to the dictionary.

To solve this, our currentpagedevice copies child dictionaries to dictionaries
allocated in local VM (and writeable). This fails if we attempt to do a
setpagedevice while the VM allocation mode is local, because the parent
dictionary is in global VM, and we then attempt to put locally allocated
children in it.

The earlier commit resolved this by forcing the VM allocation mode to local
during the course of setpagedevice, which meant that the dictionary returned
by currentpagedevice was allocated in local VM, so it was safe to put the
child dictionaries in it.

But, if we run with -dSAFER, the in gs_init.ps we do an early call to
setpagedevice in order to set /.LockSafetyParams. We do this while the VM
allocation mode is global. Note that setpagedevice stores a pointer to the
returned dictionary in the graphics state. Now, when we run the Display
PostScript startup code, it insists that all objects in the 'savedinitialgstate'
must be in global VM. But we've saved the current page deice there, and it
was allocated in local VM, which leads to an error.

So in order to satisfy the WP bug, we need to have the children allocated in
local VM. In order to call setpagedevice when the allocation mode is global
we need to create the page device dict in local VM. In order to satisfy the
DPS code, we need to allocate the page device dict in global VM.

Oops.

This commit attempts to work around this by setting the VM allocation mode
to local for the course of setpagedevice, but *not* when we are executing
the code for -dSAFER.

This is a hacky solution, and I suspect prone to failure, but probably only
if we change the startup code, which we rarely do.

No diffrences expected